♻️ refactor: 프로젝트 목록 검색 파라미터 파서를 공용 모듈로 분리
프로젝트 목록은 매칭(/matching/projects)과 지원자용(/projects) 두 경로에서
같은 필터·페이지 파라미터를 쓴다. 매칭 라우트 안에 있던 파서를 모델 계층으로
옮겨 두 화면이 같은 규칙을 보게 했다.

옮긴 로직은 그대로다. 파트 화이트리스트, 배열/단일 값 처리, 페이지 하한 1 모두
기존 동작을 유지한다.
